Brok common_test out of logger_test to avoid coupling to line numbers
