+++ logger_test.go
@@ -23,7 +23,6 @@ package zap
 import (
 	"fmt"
 	"io/ioutil"
-	"os"
 	"strings"
 	"sync"
 	"testing"
@@ -33,49 +32,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func opts(opts ...Option) []Option {
-	return opts
-}
-
-type stubbedExit struct {
-	Status *int
-}
-
-func (se *stubbedExit) Unstub() {
-	_exit = os.Exit
-}
-
-func (se *stubbedExit) AssertNoExit(t testing.TB) {
-	assert.Nil(t, se.Status, "Unexpected exit.")
-}
-
-func (se *stubbedExit) AssertStatus(t testing.TB, expected int) {
-	if assert.NotNil(t, se.Status, "Expected to exit.") {
-		assert.Equal(t, expected, *se.Status, "Unexpected exit code.")
-	}
-}
-
-func stubExit() *stubbedExit {
-	stub := &stubbedExit{}
-	_exit = func(s int) { stub.Status = &s }
-	return stub
-}
-
-func withJSONLogger(t testing.TB, enab LevelEnabler, opts []Option, f func(Logger, *testBuffer)) {
-	sink := &testBuffer{}
-	errSink := &testBuffer{}
-
-	allOpts := make([]Option, 0, 2+len(opts))
-	allOpts = append(allOpts, ErrorOutput(errSink))
-	allOpts = append(allOpts, opts...)
-	logger := New(
-		WriterFacility(newJSONEncoder(NoTime()), sink, enab),
-		allOpts...)
-
-	f(logger, sink)
-	assert.Empty(t, errSink.String(), "Expected error sink to be empty.")
-}
 
 func TestDynamicLevel(t *testing.T) {
 	lvl := DynamicLevel()
@@ -374,7 +330,7 @@ func TestLoggerAddCaller(t *testing.T) {
 	withJSONLogger(t, DebugLevel, opts(AddCaller()), func(logger Logger, buf *testBuffer) {
 		logger.Info("Callers.")
 		assert.Regexp(t,
-			`"caller":"[^"]+/logger_test.go:375","msg":"Callers\."`,
+			`"caller":"[^"]+/logger_test.go:\d+","msg":"Callers\."`,
 			buf.Stripped(), "Expected to find package name and file name in output.")
 	})
 }
@@ -383,7 +339,7 @@ func TestLoggerAddCallerSkip(t *testing.T) {
 	withJSONLogger(t, DebugLevel, opts(AddCallerSkip(1)), func(logger Logger, buf *testBuffer) {
 		logger.Info("Callers.")
 		assert.Regexp(t,
-			`"caller":"[^"]+/logger_test.go:76","msg":"Callers\."`,
+			`"caller":"[^"]+/common_test.go:\d+","msg":"Callers\."`,
 			buf.Stripped(), "Expected to find package name and file name in output.")
 	})
 }
@@ -442,16 +398,4 @@ func TestLoggerConcurrent(t *testing.T) {
 		expected := `{"level":"info","msg":"info","foo":"bar"}` + "\n"
 		assert.Equal(t, strings.Repeat(expected, 100), buf.String())
 	})
-}
-
-func runConcurrently(goroutines, iterations int, wg *sync.WaitGroup, f func()) {
-	wg.Add(goroutines)
-	for g := 0; g < goroutines; g++ {
-		go func() {
-			defer wg.Done()
-			for i := 0; i < iterations; i++ {
-				f()
-			}
-		}()
-	}
 }
